move crate and pallet into models
